Basic Implementation of Distance
Added a basic update for distance travelled and corrected a few things.
Need to address screen flicker and as ever, the display can be improved.
Also a unit conversion and perhaps lap timing mode for the stop watch.
Bluetooth to come and need the hall effect to finally arrive in the post