First Commit of Fan Drying Bench
Taking a break from pointers and keyboards. I'm now creating a simple
program that will start several fans on a drying rack depending on an
input. At the moment that input is pressure but we can use thermocouples
or a capacitence sensor. This is just the first commit to begin to
create a skeleton for the program.